+++ src/Resources/contao/languages/de/explain.php
@@ -1,0 +1,8 @@
+<?php
+
+$GLOBALS['TL_LANG']['XPL']['fideid_templates'] = array
+(
+	array('Inserttags', 'Weitere Informationen zu Inserttags finden Sie unter <a href="https://docs.contao.org/books/manual/current/de/04-inhalte-verwalten/inserttags.html" title="Contao Online-Dokumentation" target="_blank" rel="noreferrer noopener">docs.contao.org/books/manual/current/de/04-inhalte-verwalten/inserttags.html</a>.'),
+	array('colspan', '<b>Verwendbare Platzhalter</b>'),
+	array('colspan', '<b style="display:inline-block; width:200px;">##css##</b> Interne Style-Sheet-Definitionen<br><b style="display:inline-block; width:200px;">##lizenz_vorname##</b> Vorname der lizenzierten Person<br><b style="display:inline-block; width:200px;">##lizenz_nachname##</b> Nachname der lizenzierten Person<br><b style="display:inline-block; width:200px;">##lizenz_geschlecht##</b> Geschlecht (m/w/d)<br><b style="display:inline-block; width:200px;">##lizenz_art##</b> Ausbildungsgang: A, B, C-Sonderlizenz usw.<br><b style="display:inline-block; width:200px;">##lizenz_nummer##</b> Lizenznummer des DOSB, z.B. DSchB-T-C-0002146<br><b style="display:inline-block; width:200px;">##lizenz_content##</b> Inhalt aus der Textarea der E-Mail<br><b style="display:inline-block; width:200px;">##lizenz_signatur##</b> Signatur aus der E-Mail bzw. aus den System-Einstellungen'),
+);