+++ src/Resources/contao/languages/de/tl_fideid_mails.php
@@ -1,0 +1,39 @@
+<?php 
+
+/**
+ * Contao Open Source CMS
+ * Copyright (C) 2005-2014 Leo Feyer
+ *
+ */
+
+/**
+ * Backend-Modul Übersetzungen
+ */
+
+// Standardfunktionen
+$GLOBALS['TL_LANG']['tl_fideid_mails']['new']= array('Neue E-Mail', 'Neue E-Mail anlegen');
+$GLOBALS['TL_LANG']['tl_fideid_mails']['edit']= array("E-Mail %s bearbeiten", "E-Mail %s bearbeiten");
+$GLOBALS['TL_LANG']['tl_fideid_mails']['copy']= array("E-Mail %s kopieren", "E-Mail %s kopieren");
+$GLOBALS['TL_LANG']['tl_fideid_mails']['delete']= array("E-Mail %s löschen", "E-Mail %s löschen");
+$GLOBALS['TL_LANG']['tl_fideid_mails']['send']= array("E-Mail %s versenden", "E-Mail %s versenden");
+$GLOBALS['TL_LANG']['tl_fideid_mails']['show']= array("Details zur E-Mail %s anzeigen", "Details zur E-Mail %s anzeigen");
+
+// Formularfelder
+$GLOBALS['TL_LANG']['tl_fideid_mails']['text_legend'] = 'Vorlage und Text';
+$GLOBALS['TL_LANG']['tl_fideid_mails']['subject']= array('Betreff', 'Betreffzeile der E-Mail');
+$GLOBALS['TL_LANG']['tl_fideid_mails']['content']= array('Text', 'Textinhalt der E-Mail');
+
+$GLOBALS['TL_LANG']['tl_fideid_mails']['template_legend'] = 'Vorschau';
+$GLOBALS['TL_LANG']['tl_fideid_mails']['template']= array('E-Mail-Template', 'Ein Templatewechsel ändert automatisch die Vorschau');
+$GLOBALS['TL_LANG']['tl_fideid_mails']['signatur']= array('Signatur', 'Standardsignatur verwenden');
+$GLOBALS['TL_LANG']['tl_fideid_mails']['preview']= array('Vorschau', 'Vorschau mit diesem Template');
+
+$GLOBALS['TL_LANG']['tl_fideid_mails']['mail_legend'] = 'E-Mail-Optionen';
+$GLOBALS['TL_LANG']['tl_fideid_mails']['insertLizenz'] = array('Lizenzurkunde DIN A4 einfügen', 'PDF der Lizenzurkunde im Format DIN A4 als Anhang mitschicken');
+$GLOBALS['TL_LANG']['tl_fideid_mails']['insertLizenzCard'] = array('Lizenzurkunde Karte einfügen', 'PDF der Lizenzurkunde im Format Karte als Anhang mitschicken');
+$GLOBALS['TL_LANG']['tl_fideid_mails']['copyVerband']= array('E-Mail-Kopie an Landesverband', 'E-Mail-Kopie an Landesverband');
+$GLOBALS['TL_LANG']['tl_fideid_mails']['copyDSB']= array('E-Mail-Kopie an DSB', 'E-Mail-Kopie an DSB');
+
+$GLOBALS['TL_LANG']['tl_fideid_mails']['sent_state']= array('Versandstatus', 'Versandstatus');
+$GLOBALS['TL_LANG']['tl_fideid_mails']['sent_date']= array('Abgesendet am', 'Versanddatum');
+$GLOBALS['TL_LANG']['tl_fideid_mails']['sent_text']= array('Gesendeter Text', 'Gesendeter Text');