+++ src/Resources/contao/languages/de/tl_fideid_templates.php
@@ -25,7 +25,7 @@ $GLOBALS['TL_LANG']['tl_fideid_templates']['show'] = array("Details zum Template
 $GLOBALS['TL_LANG']['tl_fideid_templates']['name_legend'] = 'Name und Inhalt';
 $GLOBALS['TL_LANG']['tl_fideid_templates']['name']= array('Name', 'Name des Templates');
 $GLOBALS['TL_LANG']['tl_fideid_templates']['description']= array('Beschreibung', 'Kurze Beschreibung des Templates');
-$GLOBALS['TL_LANG']['tl_fideid_templates']['template']= array('Inhalt', 'Inhalt des Templates. Verwenden Sie den Hilfe-Link für weitere Informationen.');
+$GLOBALS['TL_LANG']['tl_fideid_templates']['template']= array('HTML-Inhalt', 'HTML-Inhalt des Templates. Verwenden Sie den Hilfe-Link für weitere Informationen.');
 
 $GLOBALS['TL_LANG']['tl_fideid_templates']['publish_legend'] = 'Aktivierung';
 $GLOBALS['TL_LANG']['tl_fideid_templates']['published']= array('Aktiv', 'Template aktivieren');
@@ -35,17 +35,13 @@ $GLOBALS['TL_LANG']['tl_fideid_templates']['default_template']=
 '<!DOCTYPE html PUBLIC "-//W3C//DTD HTML 3.2//EN">
 <html>
 <head>
-	<meta http-equiv="Content-Type" content="text/html; charset=<?= $this->charset ?>">
+	<meta http-equiv="Content-Type" content="text/html; charset=utf-8">
 	<meta name="Generator" content="Contao Open Source CMS">
-	<title><?= $this->title ?></title>
+	<title>##subject##</title>
 </head>
 <body>
 
-	<?= $this->content ?>
-
-	<?php if($this->signatur): ?>
-		<?= $this->signatur ?>
-	<?php endif; ?>
+	##content##
 
 </body>
 </html>';