+++ src/Resources/contao/languages/de/tl_settings.php
@@ -1,0 +1,12 @@
+<?php
+
+/**
+ * Legende
+ */
+$GLOBALS['TL_LANG']['tl_settings']['fideidverwaltung_legend']       = 'FIDE-ID-Nummern';
+
+/**
+ * Felder
+ */
+$GLOBALS['TL_LANG']['tl_settings']['fideidverwaltung_absender'] = array('Absender für E-Mail-Versand', 'Absender (Name und E-Mail) für den E-Mail-Versand');
+$GLOBALS['TL_LANG']['tl_settings']['fideidverwaltung_mailsignatur'] = array('Signatur für E-Mail-Versand', 'Signatur für den E-Mail-Versand');